nvim: Add the Neogit plugin
In this commit, I added the Neogit plugin as I find it useful to view
the diffview whilst writing commits.
The plugin is lazy-loaded and only is loaded when the `:Neogit` command
is ran so the startup overhead should be minimal.

neogit: dofile the git integration when setup is called