nvim: replace ibl w blink.indent
Honestly, indent-blankline.nvim is quick enough to not feel any lag but
I would like to try blink.indent for some time to see if it has any
differences or not.

As of right now, I'm lazy-loading the plugin as I don't need it on
startup but, as per it's README, the end-user doesn't need to manually
lazy-load so I might just give up on lazy-loading that as a whole.

^^ Of course that's IF I go around keeping this plugin. It's a smaller,
newer and less popular plugin so it might have some rough edges. As of
writing, I really don't have any issues so far so that's that, I guess.